ref: asset details refactored
