+++ src/components/AssetDetails/AssetDetails.module.css
@@ -7,34 +7,6 @@
   gap: 24px;
 }
 
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
-.inputs {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.actions {
-  display: flex;
-  gap: 24px;
-}
-
-.actions button {
-  flex: 1;
-}
-
-.errors {
-  font-size: 12px;
-  color: var(--error);
-  margin-top: -24px;
-  align-self: flex-start;
-}
-
 .tradeBtn {
   width: 100%;
 }
