+++ src/components/TradeModal/TradeModal.module.css
@@ -1,0 +1,27 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
+
+.inputs {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+
+.actions {
+  display: flex;
+  gap: 24px;
+}
+
+.actions button {
+  flex: 1;
+}
+
+.errors {
+  font-size: 12px;
+  color: var(--error);
+  margin-top: -24px;
+  align-self: flex-start;
+}
