+++ src/types/index.ts
@@ -1,0 +1,6 @@
+import { PropsWithChildren } from 'react';
+
+export type ModalProps = PropsWithChildren & {
+  isOpen: boolean;
+  onClose?: () => void;
+};
